// SmartBalance/Version 1.0.0/Code/Frontend/mersal-accounting-app/src/app/BusinessFeatures/AccountsTree/Models/AccountTree.model.ts
export class accountTree {
    code: string;
    fullCode: string;
    id: number;
    nameAR: string;
    nameEN: string;
    
    descriptionAr: string;
    descriptionEn: string;

    accountTypeId: string;
    date: Date;
    openingCredit: number;
    isDebt: boolean;
    groupId: number;
    categoryId: number;
    parentAccountChartId: number;
    isFinalNode: boolean;
    currencyId: any;
    currentDebitBalance: any;
    currentCreditBalance: any;
    isCreditAccount: boolean;
    branchId: any;
}